Made hurl file for logging in and retrieving token
login.hurl is hurl file for logging in and retrieving token using hurl

secrets.template is template file for secrets and should never be edited
instead a copy named secrets.env should be created and edited.
secret.env is in .gitignore and will never be committed, making it
safer.

get_token.sh is a shorthand for running the login.hurl file with
secrets.env as variables.